+++ packages/live-state/src/server/adapters/express/index.ts
@@ -1,7 +1,7 @@
 import { Application } from "express-ws";
 import { AnyRouter, Server } from "../../";
 import { httpTransportLayer } from "../../transport-layers/http";
-import { webSocketAdapter } from "../../web-socket";
+import { webSocketAdapter } from "../../transport-layers/web-socket";
 import { convertRequest } from "./convert-request";
 
 export const expressAdapter = (
