+++ packages/live-state/src/server/index.ts
@@ -6,7 +6,6 @@ import { Storage } from "./storage";
 export * from "./adapters/express";
 export * from "./router";
 export * from "./storage";
-export * from "./web-socket";
 
 export type Request<TInput = any> = {
   headers: Record<string, string>;
